Remove redundant getter from config. Move whitelist into app folder. Add the ytdl pool size to the config settings. Switch from plain print statements to a logging library. Add some additional logging for clarity. Add new python dependency to the debian install script.
