+++ example.ini
@@ -1,3 +1,6 @@
 [server]
 host=localhost
 port=8420
+
+[ytdl]
+pool_size=10
